feature: delete old dynamic groupping stuff (named userapi)
