+++ src/apimachinery/hostserver/apis.go
@@ -438,90 +438,6 @@ func (hs *hostServer) UpdateHostPropertyBatch(ctx context.Context, h http.Header
 	return
 }
 
-func (hs *hostServer) AddUserCustomQuery(ctx context.Context, h http.Header, dat map[string]interface{}) (resp *metadata.Response, err error) {
-	resp = new(metadata.Response)
-	subPath := "/userapi"
-
-	err = hs.client.Post().
-		WithContext(ctx).
-		Body(dat).
-		SubResourcef(subPath).
-		WithHeaders(h).
-		Do().
-		Into(resp)
-	return
-}
-
-func (hs *hostServer) UpdateUserCustomQuery(ctx context.Context, businessID string, id string, h http.Header, dat map[string]interface{}) (resp *metadata.Response, err error) {
-	resp = new(metadata.Response)
-	subPath := "/userapi/%s/%s"
-
-	err = hs.client.Put().
-		WithContext(ctx).
-		Body(dat).
-		SubResourcef(subPath, businessID, id).
-		WithHeaders(h).
-		Do().
-		Into(resp)
-	return
-}
-
-func (hs *hostServer) DeleteUserCustomQuery(ctx context.Context, businessID string, id string, h http.Header) (resp *metadata.Response, err error) {
-	resp = new(metadata.Response)
-	subPath := "/userapi/%s/%s"
-
-	err = hs.client.Delete().
-		WithContext(ctx).
-		Body(nil).
-		SubResourcef(subPath, businessID, id).
-		WithHeaders(h).
-		Do().
-		Into(resp)
-	return
-}
-
-func (hs *hostServer) GetUserCustomQuery(ctx context.Context, businessID string, h http.Header, dat *metadata.QueryInput) (resp *metadata.Response, err error) {
-	resp = new(metadata.Response)
-	subPath := "/userapi/search/%s"
-
-	err = hs.client.Post().
-		WithContext(ctx).
-		Body(dat).
-		SubResourcef(subPath, businessID).
-		WithHeaders(h).
-		Do().
-		Into(resp)
-	return
-}
-
-func (hs *hostServer) GetUserCustomQueryDetail(ctx context.Context, businessID string, id string, h http.Header) (resp *metadata.UserCustomQueryDetailResult, err error) {
-	resp = new(metadata.UserCustomQueryDetailResult)
-	subPath := "/userapi/detail/%s/%s"
-
-	err = hs.client.Get().
-		WithContext(ctx).
-		Body(nil).
-		SubResourcef(subPath, businessID, id).
-		WithHeaders(h).
-		Do().
-		Into(resp)
-	return
-}
-
-func (hs *hostServer) GetUserCustomQueryResult(ctx context.Context, businessID, id, start, limit string, h http.Header) (resp *metadata.Response, err error) {
-	resp = new(metadata.Response)
-	subPath := "/userapi/data/%s/%s/%s/%s"
-
-	err = hs.client.Get().
-		WithContext(ctx).
-		Body(nil).
-		SubResourcef(subPath, businessID, id, start, limit).
-		WithHeaders(h).
-		Do().
-		Into(resp)
-	return
-}
-
 // CreateDynamicGroup is dynamic group create action api machinery.
 func (hs *hostServer) CreateDynamicGroup(ctx context.Context, header http.Header,
 	data map[string]interface{}) (resp *metadata.IDResult, err error) {
