+++ src/common/metadata/hostserver.go
@@ -50,11 +50,6 @@ type HostSnapBatchResult struct {
 	Data     []map[string]interface{} `json:"data"`
 }
 
-type UserCustomQueryDetailResult struct {
-	BaseResp `json:",inline"`
-	Data     map[string]interface{} `json:"data"`
-}
-
 type HostInputType string
 
 const (
