+++ src/scene_server/admin_server/upgrader/y3.9.202010131456/pkg.go
@@ -14,10 +14,23 @@ package y3_9_202010131456
 
 import (
 	"context"
+	"time"
 
 	"configcenter/src/scene_server/admin_server/upgrader"
 	"configcenter/src/storage/dal"
 )
+
+// UserConfig is old dynamic groupping metadata struct, now only used in migrate.
+type UserConfig struct {
+	Info       string    `json:"info" bson:"info"`
+	Name       string    `json:"name" bson:"name"`
+	ID         string    `json:"id" bson:"id"`
+	CreateTime time.Time `json:"create_time" bson:"create_time"`
+	UpdateTime time.Time `json:"last_time" bson:"last_time"`
+	AppID      int64     `json:"bk_biz_id" bson:"bk_biz_id"`
+	CreateUser string    `json:"create_user" bson:"create_user"`
+	ModifyUser string    `json:"modify_user" bson:"modify_user"`
+}
 
 func init() {
 	upgrader.RegistUpgrader("y3.9.202010131456", upgrade)
